-added a separate method to check permissions -implemented method to requestPermissions -implemented method to get coordinates -changed retrofit interface parameters (lat,lon) from Integer to Double

+++ app/src/main/java/ru/makproductions/androidlevel2/OpenWeatherRetrofit.java
@@ -8,5 +8,5 @@ public interface OpenWeatherRetrofit {
     @GET("data/2.5/weather")
     Call<WeatherMap> loadWeather(@Query("q") String cityName, @Query("appid") String appId, @Query("units") String units);
     @GET("data/2.5/weather")
-    Call<WeatherMap> loadWeatherByCoordinates(@Query("lat") Integer latitude,@Query("lon") Integer longtitude, @Query("appid") String appId, @Query("units") String units);
+    Call<WeatherMap> loadWeatherByCoordinates(@Query("lat") Double latitude,@Query("lon") Double longtitude, @Query("appid") String appId, @Query("units") String units);
 }
